Move template back to header file
